+++ azimuth/modules/base_classes/dask_module.py
@@ -1,4 +1,7 @@
-# Copyright (c) 2021 ServiceNow Inc. All rights reserved.
+# Copyright ServiceNow, Inc. 2021 – 2022
+# This source code is licensed under the Apache 2.0 license found in the LICENSE file
+# in the root directory of this source tree.
+
 import abc
 import os
 import threading